RAFT raft.py cot answers now checkpointed
